requirements.txt
absl-py==0.8.1
acora==2.2
asn1crypto==0.24.0
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
bcolz==1.2.1
beautifulsoup4==4.8.0
bleach==3.0.2
blessed==1.15.0
boto==2.49.0
boto3==1.9.58
botocore==1.12.58
bz2file==0.98
cachetools==3.0.0
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
codegen==1.0
colorama==0.4.1
colormath==3.0.0
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==2.7
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.2
Cython==0.29.13
cytoolz==0.9.0.1
dask==1.0.0
data-science-utils==0.0.0
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.8.2
distorm3==3.4.1
distro==1.4.0
dlib==19.18.0
docutils==0.14
easydict==1.9
elasticsearch==7.1.0
en-core-web-sm==2.0.0
entrypoints==0.2.3
enum34==1.1.6
filelock==3.0.10
filemagic==1.6
Flask==1.0.2
Flask-Cors==3.0.7
funcsigs==1.0.2
future==0.17.1
gapic-google-cloud-datastore-v1==0.90.4
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gast==0.2.2
gensim==3.6.0
gevent==1.4.0
google-api-core==1.5.2
google-api-python-client==1.7.5
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
google-cloud==0.34.0
google-cloud-bigquery==1.7.0
google-cloud-bigquery-datatransfer==0.2.0
google-cloud-bigtable==0.31.1
google-cloud-container==0.1.1
google-cloud-core==0.29.0
google-cloud-datastore==1.7.1
google-cloud-dns==0.29.0
google-cloud-error-reporting==0.30.0
google-cloud-firestore==0.30.0
google-cloud-language==1.1.0
google-cloud-logging==1.8.0
google-cloud-monitoring==0.31.0
google-cloud-pubsub==0.39.0
google-cloud-resource-manager==0.28.1
google-cloud-runtimeconfig==0.28.1
google-cloud-spanner==1.6.1
google-cloud-speech==0.36.0
google-cloud-storage==1.13.0
google-cloud-trace==0.19.0
google-cloud-translate==1.3.1
google-cloud-videointelligence==1.6.0
google-cloud-vision==0.35.1
google-gax==0.16.0
google-images-download==2.8.0
google-pasta==0.1.7
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
GPUtil==1.4.0
greenlet==0.4.15
grpc-google-iam-v1==0.11.4
grpcio==1.16.1
h2o==3.22.0.2
h5py==2.8.0
hashids==1.2.0
html5lib==1.0.1
httplib2==0.12.0
humanize==0.5.1
hyperlink==19.0.0
ibm-cloud-sdk-core==1.0.0
ibm-watson==4.0.1
idna==2.7
imgkit==1.0.2
imutils==0.5.3
incremental==17.5.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
joblib==0.13.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
Keras==2.1.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-resnet==0.1.0
keras-retinanet==0.5.1
keras-tqdm==2.0.1
kiwisolver==1.0.1
lightgbm==2.2.2
llvmlite==0.26.0
lxml==4.4.1
lz4==2.2.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
mock==3.0.5
more-itertools==4.3.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
nltk==3.4
notebook==5.7.2
numpy==1.17.2
oauth2client==4.1.3
oauthlib==2.1.0
olefile==0.46
opencv-python==3.4.4.19
opt-einsum==3.1.0
pandas==0.23.4
pandocfilters==1.4.2
parsel==1.5.2
parso==0.3.1
patsy==0.5.1
pdfminer.six==20181108
pdfplumber==0.5.12
pefile==2019.4.18
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
plac==0.9.6
pluggy==0.8.0
ply==3.11
preshed==2.0.1
progressbar2==3.46.1
prometheus-client==0.4.2
promise==2.2.1
prompt-toolkit==2.0.7
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.6.1
psutil==5.4.8
ptyprocess==0.6.0
py==1.7.0
py2neo==1.6.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.8.2
PyDispatcher==2.0.5
pydot==1.4.1
pyelftools==0.25
Pygments==2.3.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymongo==3.9.0
PyMuPDF==1.14.16
pyOpenSSL==19.0.0
pyparsing==2.3.0
pyprof2calltree==1.4.4
pytest==4.0.1
python-dateutil==2.7.5
python-utils==2.3.0
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
queuelib==1.5.0
regex==2018.1.10
rekall==1.0.3
requests==2.20.1
requests-oauthlib==1.0.0
rsa==4.0
s3transfer==0.1.13
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.1.0
Scrapy==1.7.3
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
setproctitle==1.1.10
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
smart-open==1.7.1
sortedcontainers==2.1.0
soupsieve==1.9.4
spacy==2.0.18
ssdeep===2.9-0.3
statsmodels==0.9.0
tabula-py==1.4.1
tabulate==0.8.2
tensorboard==1.11.0
tensorflow==1.11.0
tensorflow-datasets==1.2.0
tensorflow-estimator==1.14.0
tensorflow-metadata==0.14.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==1.0.4
thinc==6.12.1
thrift==0.11.0
toolz==0.9.0
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
Twisted==19.7.0
ujson==1.35
unicodecsv==0.14.1
UNKNOWN==0.0.0
uritemplate==3.0.0
urllib3==1.24.1
w3lib==1.21.0
Wand==0.5.4
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.48.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
workbench==0.3.6
workbench-cli==0.3.3
wrapt==1.11.2
yara==1.7.7
zerorpc==0.6.3
zope.interface==4.6.0
